Switch to BETA 2
French version / Correction de bug :
-Corrige un dépassement de capacité dans la calcul de durée quand les vidéos à générer sont longues >30mn
-Corrige une erreur de perte de l'orientation d'image lors de l'activation de filtre de transformation
-Change le synchro d'affichage au demarrage de l'application (ajoute 500 ms)

English version / Bug fix :
-Correct an overflow in calculation of duration when videos to be generated are long > 30mn
-Correct an error of loss of the image orientation during the activation of transformation filter
-Change display synchronization at application startup (add 500 ms)

French version / Nouvelles fonctionnalités :
-Ajoute une option dans la boite de dialog Option du programme pour activer le resampler rapide pendant la preview
-Ajoute une option dans la boite de dialog Option du programme pour sélectionner le nom par defaut de la vidéo à générer

English version / New features :
-Add an option in configuration dialog to activate quick resampler during preview
-Add an option in configuration dialog to default name for rendered video

git-svn-id: file:///media/werkstatt/064f12d4-5f27-4839-9537-9ffb7939d236/ffDiaporama-Entwicklung-2023/ffdiaporama-svn/trunk@213 f803d417-eb43-4c80-8848-1a7435ce69a7

+++ ffDiaporama/DlgRenderVideo.h
@@ -42,10 +42,10 @@ public:
     int             VideoCodecIndex;                // Index of video codec
     int             AudioCodecIndex;                // Index of audio codec
     double          FPS;
-    int             NbrFrame;                       // Number of frame to generate
+    qlonglong       NbrFrame;                       // Number of frame to generate
     QTime           StartTime;                      // Time the process start
     QTime           LastCheckTime;                  // Last time the loop start
-    int             RenderedFrame;
+    qlonglong       RenderedFrame;
     bool            IsDestFileOpen;                 // true if encoding is started
 
     QString         OutputFileName;                 // filename with path for the last rendering file
